requirements.txt
PyMuPDF
Pillow
numpy
opencv-python
torch
transformers